added new configurations to resolve critical issues on SonarQube

+++ src/test/java/com/seguridad/seguridad_calidad_back/model/RecipeCommentTest.java
@@ -11,7 +11,7 @@ public class RecipeCommentTest {
         RecipeComment recipeComment = new RecipeComment();
         long id = 1L;
         int idReceta = 101;
-        int idUsuario = 202;
+        Long idUsuario = 202L;
         String comentario = "Delicious recipe!";
 
         // Act
